+++ changelog.md
@@ -3,6 +3,7 @@ v3.2.1
 * Fixed client crash with Enderport enchantments on multiplayer servers
 * Fixed Runic Grindstone to remove runes even from unenchanted items
 * Removed default affinity definitions from chainmail armor
+* Changed Runecarving Table slots to be filtered by tags instead of hard-coded items
 
 v3.2.0
 ------